Break it into CLI and support functions (inside python/ folder) so you can also do that interactively in a notebook if needed
Example Call to CLI:
```
python performance_test.py --urls https://openstudio-ci-builds.s3-us-west-2.amazonaws.com/develop/OpenStudio-3.3.1-alpha%2Bf3fba80d69-Darwin.tar.gz https://openstudio-ci-builds.s3-us-west-2.amazonaws.com/develop/OpenStudio-3.3.1-alpha%2B44a7660fbb-Darwin.tar.gz --number-runs 2 --test-files baseline_sys01.rb zone_hvac.rb airloop_and_zonehvac.rb
```
